+++ src/main/java/io/github/sunday/devfolio/controller/portfolio/PortfolioController.java
@@ -8,18 +8,13 @@ import io.github.sunday.devfolio.service.portfolio.PortfolioLikeService;
 import io.github.sunday.devfolio.service.portfolio.PortfolioService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.jsoup.Jsoup;
-import org.jsoup.safety.Safelist;
-import org.springframework.beans.propertyeditors.StringTrimmerEditor;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.ObjectError;
-import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
 
-import java.beans.PropertyEditorSupport;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -33,37 +28,6 @@ public class PortfolioController {
     private final PortfolioService portfolioService;
     private final PortfolioCategoryService portfolioCategoryService;
     private final PortfolioLikeService portfolioLikeService;
-
-    /**
-     * 포트폴리오 검색 요청이 들어올 때 DTO 내의 String에서 script를 제거
-     */
-    @InitBinder("searchRequestDto")
-    public void initBinder(WebDataBinder binder) {
-        // keyword와 category의 sanitize 수행
-        binder.registerCustomEditor(String.class, new StringTrimmerEditor(true) {
-            @Override
-            public void setAsText(String text) {
-                if (text != null) {
-                    String safeText = Jsoup.clean(text, Safelist.basic());
-                    super.setAsText(safeText.trim());
-                } else {
-                    super.setValue(null);
-                }
-            }
-        });
-
-        // sort 유효성 검증
-        binder.registerCustomEditor(PortfolioSort.class, new PropertyEditorSupport() {
-            @Override
-            public void setAsText(String text) throws IllegalArgumentException {
-                PortfolioSort sort = PortfolioSort.fromName(text);
-                if (sort == null) {
-                    sort = PortfolioSort.UPDATED_AT;
-                }
-                setValue(sort);
-            }
-        });
-    }
 
     /**
      * 포트폴리오 메인 페이지 출력
