+++ src/main/java/io/github/sunday/devfolio/controller/portfolio/PortfolioRestController.java
@@ -3,22 +3,16 @@ package io.github.sunday.devfolio.controller.portfolio;
 import io.github.sunday.devfolio.config.CustomUserDetails;
 import io.github.sunday.devfolio.dto.portfolio.PortfolioListDto;
 import io.github.sunday.devfolio.dto.portfolio.PortfolioSearchRequestDto;
-import io.github.sunday.devfolio.enums.PortfolioSort;
 import io.github.sunday.devfolio.service.portfolio.PortfolioLikeService;
 import io.github.sunday.devfolio.service.portfolio.PortfolioService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.jsoup.Jsoup;
-import org.jsoup.safety.Safelist;
-import org.springframework.beans.propertyeditors.StringTrimmerEditor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.validation.BindingResult;
 import org.springframework.validation.FieldError;
-import org.springframework.web.bind.WebDataBinder;
 import org.springframework.web.bind.annotation.*;
 
-import java.beans.PropertyEditorSupport;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -35,42 +29,12 @@ public class PortfolioRestController {
     private final PortfolioLikeService portfolioLikeService;
 
     /**
-     * 포트폴리오 검색 요청이 들어올 때 DTO 내의 String에서 script를 제거
-     */
-    @InitBinder("requestDto")
-    public void initBinder(WebDataBinder binder) {
-        binder.registerCustomEditor(String.class, new StringTrimmerEditor(true) {
-            @Override
-            public void setAsText(String text) {
-                if (text != null) {
-                    String safeText = Jsoup.clean(text, Safelist.basic());
-                    super.setAsText(safeText.trim());
-                } else {
-                    super.setValue(null);
-                }
-            }
-        });
-
-        // sort 유효성 검증
-        binder.registerCustomEditor(PortfolioSort.class, new PropertyEditorSupport() {
-            @Override
-            public void setAsText(String text) throws IllegalArgumentException {
-                PortfolioSort sort = PortfolioSort.fromName(text);
-                if (sort == null) {
-                    sort = PortfolioSort.UPDATED_AT;
-                }
-                setValue(sort);
-            }
-        });
-    }
-
-    /**
      * 포트폴리오 목록 출력 API
      * 무한 스크롤에 적용되는 정보 요청 API
      */
     @GetMapping("/list")
     public ResponseEntity<?> list(
-            @Valid @ModelAttribute("requestDto") PortfolioSearchRequestDto requestDto,
+            @Valid @ModelAttribute("searchRequestDto") PortfolioSearchRequestDto requestDto,
             BindingResult bindingResult
     ) {
         if (bindingResult.hasErrors()) {
