+++ src/main/resources/templates/portfolio/portfolio_detail.html
@@ -59,7 +59,8 @@
                     </div>
                 </div>
                 <div class="portfolio-thumbnail">
-                    <img th:src="${portfolio.thumbnailUrl == null or #strings.isEmpty(portfolio.thumbnailUrl)} ? @{/assets/images/portfolio-default-thumbnail.png}" th:alt="*{title}">
+                    <img th:src="${portfolio.thumbnailUrl == null or #strings.isEmpty(portfolio.thumbnailUrl)}
+                    ? @{/assets/images/portfolio-default-thumbnail.png} : ${portfolio.thumbnailUrl}" th:alt="*{title}">
                 </div>
                 <div th:utext="*{description}" class="description"></div>
                 <th:block th:if="${loggedIn != null && #strings.equals(loggedIn, true)}">
